requirements.txt
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2018.11.29
chardet==3.0.4
Django==2.1.4
djangorestframework==3.9.0
idna==2.7
pytz==2018.7
requests==2.20.1
urllib3==1.24.1
